Fix: optimize code
